server: moved github helpers to libgh package
